chore: address PR comments, refactor code
