+++ src/components/Nodes/BarGraphNode/BarGraphNode.tsx
@@ -1,15 +1,22 @@
-import { ResponsivePie as Pie } from '@nivo/pie'
+import { ResponsiveBar as Bar } from '@nivo/bar'
+import {
+  ChangeEvent,
+  useCallback,
+  useContext,
+  useEffect,
+  useState,
+} from 'react'
 
 import { v4 as uuid } from 'uuid'
 
+import { config } from './config'
 import { ThemeContext } from 'src/contexts'
 import { IconButton, GraphNodeWrapper, EditGraphPanel } from 'src/components'
-import { lightTheme, darkTheme } from 'src/components/BarGraphNode/theme'
-
-import styles from './PieGraphNode.module.scss'
+import { lightTheme, darkTheme } from 'src/components/Nodes/theme'
+import styles from 'src/components/Nodes/Nodes.module.scss'
 import { EditableGraphData, NodeData } from 'src/types/data/base'
 import { edit, save } from 'src/components/IconButton/icons'
-import { useCallback, useContext, useEffect, useState } from 'react'
+import classNames from 'classnames'
 
 interface Props {
   nodeData: NodeData
@@ -18,12 +25,14 @@ interface Props {
   updateNode?: (data: NodeData) => void
 }
 
-export const PieGraphNode: React.FC<Props> = (props: Props) => {
-  const { nodeData, editable = false, updateNode } = props
-  const { title, description, data } = nodeData
+export const BarGraphNode: React.FC<Props> = (props: Props) => {
   const { theme } = useContext(ThemeContext)
+  const { nodeData, editable = false, updateNode } = props
+  const { title, description, legend, data } = nodeData
 
+  const [editableLegend, setEditableLegend] = useState(legend)
   const [editActive, setEditActive] = useState(false)
+
   const [graphData, setGraphData] = useState<EditableGraphData>({})
 
   const onTitleSave = useCallback(
@@ -48,6 +57,10 @@ export const PieGraphNode: React.FC<Props> = (props: Props) => {
     )
   }, [nodeData])
 
+  const handleLegendChange = useCallback((e: ChangeEvent<HTMLInputElement>) => {
+    setEditableLegend(e.target.value)
+  }, [])
+
   return (
     <GraphNodeWrapper
       title={title}
@@ -57,57 +70,29 @@ export const PieGraphNode: React.FC<Props> = (props: Props) => {
       editable={editable}
     >
       <div className={styles.container}>
-        <div className={editActive ? styles.editable : styles.graph}>
-          <Pie
+        <div
+          className={classNames(styles.graph, {
+            [styles.active]: editActive,
+          })}
+        >
+          <Bar
             data={
               editable
                 ? Object.values(graphData).map(([key, value]) => ({
-                    id: key,
                     label: key,
                     value,
                   }))
                 : Object.entries(data).map(([key, value]) => ({
-                    id: key,
                     label: key,
                     value,
                   }))
             }
-            margin={{ top: 10, right: 80, bottom: 70, left: 80 }}
-            innerRadius={0.5}
-            padAngle={0.7}
-            cornerRadius={3}
-            activeOuterRadiusOffset={8}
-            borderWidth={1}
-            borderColor={{
-              from: 'color',
-              modifiers: [['darker', 0.2]],
+            {...config}
+            axisTop={{
+              legend: editActive ? editableLegend : legend,
+              tickValues: [],
+              legendPosition: 'middle',
             }}
-            arcLinkLabel="id"
-            arcLinkLabelsSkipAngle={10}
-            arcLinkLabelsThickness={2}
-            arcLinkLabelsColor={{ from: 'color' }}
-            arcLabelsSkipAngle={10}
-            arcLabelsTextColor={{
-              from: 'color',
-              modifiers: [['darker', 2]],
-            }}
-            colors={{ scheme: 'green_blue' }}
-            legends={[
-              {
-                anchor: 'bottom',
-                direction: 'row',
-                justify: false,
-                translateX: 0,
-                translateY: 56,
-                itemsSpacing: 0,
-                itemWidth: 100,
-                itemHeight: 18,
-                itemDirection: 'left-to-right',
-                itemOpacity: 1,
-                symbolSize: 18,
-                symbolShape: 'circle',
-              },
-            ]}
             theme={theme == 'dark' ? darkTheme : lightTheme}
           />
         </div>
@@ -122,13 +107,19 @@ export const PieGraphNode: React.FC<Props> = (props: Props) => {
         )}
         {editable && updateNode && editActive && (
           <div className={styles.panel}>
-            <EditGraphPanel data={graphData} handleDataChange={setGraphData} />
+            <EditGraphPanel
+              data={graphData}
+              legend={editableLegend}
+              handleLegendChange={handleLegendChange}
+              handleDataChange={setGraphData}
+            />
             <IconButton
               onClick={async () => {
                 setEditActive(false)
                 updateNode({
                   ...nodeData,
                   data: Object.fromEntries(Object.values(graphData)),
+                  legend: editableLegend,
                 })
               }}
             >
