+++ src/components/Nodes/BarGraphNode/index.ts
