+++ src/components/Nodes/Nodes.module.scss
@@ -2,13 +2,10 @@
 
 .container {
   @include row;
-  height: 60%;
   width: 100%;
-  @include themed-styles {
-    border-bottom: 2px solid rgba(themed-colour('text-muted'), 0.5);
-  }
-
+  flex: 1;
   position: relative;
+
   & > button {
     position: absolute;
     right: 0;
@@ -18,10 +15,10 @@
 .graph {
   height: 100%;
   width: 100%;
+  position: absolute;
 }
 
-.editable {
-  height: 100%;
+.active {
   width: 60%;
 }
 
@@ -29,4 +26,7 @@
   @include row;
   height: 100%;
   width: 40%;
+  position: absolute;
+  overflow-y: scroll;
+  right: 40px;
 }
