+++ src/components/Nodes/index.ts
@@ -1,0 +1,2 @@
+export * from './BarGraphNode'
+export * from './PieGraphNode'
