+++ src/components/index.ts
@@ -10,10 +10,9 @@ export * from './DropdownMenu'
 export * from './SearchBar'
 export * from './ProjectListItem'
 export * from './RootNode'
-export * from './BarGraphNode'
+export * from './Nodes'
 export * from './IconButton'
 export * from './EditableHeading'
 export * from './EditableDescription'
 export * from './EditGraphPanel'
 export * from './LoadingSpinner'
-export * from './PieGraphNode'
