+++ src/styles/_theme.scss
@@ -31,6 +31,7 @@ $themes: (
     button-selected-secondary: #555558,
     button-new-project: #172436,
     shadow-primary: #0000002c,
+    border: #66a5a9ae,
   ),
   light: (
     base: #ffffff,
@@ -64,6 +65,7 @@ $themes: (
     button-selected-secondary: #747477,
     button-new-project: white,
     shadow-primary: #0000002c,
+    border: #807f7f81,
   ),
 );
